Revert "refactor assembly conflict checker"
This reverts commit bbe34d80f64aa6cfb6fb711e5629ca32479b863a.
