+++ src/DynamoApplications/Properties/Resources.resx
@@ -117,4 +117,16 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="MismatchedAssemblyList" xml:space="preserve">
+    <value>It is likely one of the following assemblies loaded the incompatible version:</value>
+  </data>
+  <data name="MismatchedAssemblyVersion" xml:space="preserve">
+    <value>While loading assembly {0}, Dynamo detected that the dependency {1} was already loaded with an incompatible version. It is likely that another Revit Addin has loaded this assembly, please try uninstalling other Addins, and starting Dynamo again. Dynamo may be unstable in this state.</value>
+  </data>
+  <data name="MismatchedAssemblyVersionShortMessage" xml:space="preserve">
+    <value>Dynamo has detected a conflict with a dependency and may be unstable. If any issues are detected, please uninstall addins or packages and contact the provider.</value>
+  </data>
+  <data name="MismatchedAssemblyVersionTitle" xml:space="preserve">
+    <value>Error Loading A Dependency</value>
+  </data>
 </root>
