+++ src/DynamoApplications/StartupUtils.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Globalization;
 using System.IO;
@@ -7,11 +8,11 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 using System.Threading;
-using Dynamo.Core;
 using Dynamo.Interfaces;
 using Dynamo.Models;
 using Dynamo.Scheduler;
 using Dynamo.Updates;
+using DynamoApplications.Properties;
 using DynamoShapeManager;
 using Microsoft.Win32;
 using NDesk.Options;
@@ -297,15 +298,77 @@ namespace Dynamo.Applications
         }
 
         /// <summary>
+        /// The white list of dependencies to be ignored.
+        /// </summary>
+        private static String[] assemblyNamesToIgnore = { "Newtonsoft.Json" };
+
+        /// <summary>
         /// Checks that an assembly does not have any dependencies that have already been loaded into the 
         /// appDomain with an incompatible to the one Dynamo requires.
         /// </summary>
         /// <param name="assembly"></param>
         /// <returns>returns a list of fileLoad exceptions - if the list is empty no mismatched assemblies were encountered </returns>
-        [Obsolete("Method will be deprecated in Dynamo 3.0, please use Dynamo.Core.AssemblyConflicNotifier instead.")]
         public static List<Exception> CheckAssemblyForVersionMismatches(Assembly assembly)
         {
-            return AssemblyConflictNotifier.GetVersionMismatchedReferencesInAppDomain(assembly, new string[] { }).ToList();
+            return GetVersionMismatchedReferencesInAppDomain(assembly, assemblyNamesToIgnore);
         }
+
+        /// <summary>
+        /// Handler for an assembly load event into a host's appdomain - we need to make sure
+        /// that another addin or package has not loaded another version of a .dll that we require.
+        /// If this happens Dynamo will most likely crash. We should alert the user they
+        /// have an incompatible addin/package installed.. this is only called if the host calls or
+        /// subscribes to it during AppDomain.AssemblyLoad event.
+        /// 
+        private static List<Exception> GetVersionMismatchedReferencesInAppDomain(Assembly assembly, String[] assemblyNamesToIgnore)
+        {
+            // Get all assemblies that are currently loaded into the appdomain.
+            var loadedAssemblies = AppDomain.CurrentDomain.GetAssemblies().ToList();
+            // Ignore some assemblies(Revit assemblies) that we know work and have changed their version number format or do not align
+            // with semantic versioning.
+
+            var loadedAssemblyNames = loadedAssemblies.Select(assem => assem.GetName()).ToList();
+            loadedAssemblyNames.RemoveAll(assemblyName =>assemblyNamesToIgnore.Contains(assemblyName.Name));
+
+            //build dict- ignore those with duplicate names.
+            var loadedAssemblyDict = loadedAssemblyNames.GroupBy(assm => assm.Name).ToDictionary(g => g.Key, g => g.First());
+
+            var output = new List<Exception>();
+
+            foreach (var currentReferencedAssembly in assembly.GetReferencedAssemblies().Concat(new AssemblyName[] { assembly.GetName() }))
+            {
+                if (loadedAssemblyDict.ContainsKey(currentReferencedAssembly.Name))
+                {
+                    //if the dll is already loaded, then check that our required version is not greater than the currently loaded one.
+                    var loadedAssembly = loadedAssemblyDict[currentReferencedAssembly.Name];
+                    if (currentReferencedAssembly.Version.Major > loadedAssembly.Version.Major)
+                    {
+                        //there must exist a loaded assembly which references the newer version of the assembly which we require - lets find it:
+
+                        var referencingNewerVersions = new List<AssemblyName>();
+                        foreach(var originalLoadedAssembly in loadedAssemblies )
+                        {
+                            foreach(var refedAssembly in originalLoadedAssembly.GetReferencedAssemblies())
+                            {
+                                //if the version matches then this is one our guys
+                                if(refedAssembly.Version == loadedAssembly.Version)
+                                {
+                                    referencingNewerVersions.Add(originalLoadedAssembly.GetName());
+                                }
+                            }
+                        }
+
+                        output.Add(new FileLoadException(
+                            string.Format(Resources.MismatchedAssemblyVersion, assembly.FullName, currentReferencedAssembly.FullName)
+                            + Environment.NewLine + Resources.MismatchedAssemblyList + Environment.NewLine +
+                            String.Join(", ", referencingNewerVersions.Select(x => x.Name).Distinct().ToArray())));
+                    }
+                }
+            }
+            return output;
+        }
+
+
+
     }
 }
