+++ src/DynamoCore/DynamoCore.csproj
@@ -102,7 +102,6 @@ limitations under the License.
       <Link>Properties\AssemblySharedInfo.cs</Link>
     </Compile>
     <Compile Include="Configuration\ExecutionSession.cs" />
-    <Compile Include="Core\AssemblyConflictNotifier.cs" />
     <Compile Include="Core\CrashPromptArgs.cs" />
     <Compile Include="Configuration\DebugSettings.cs" />
     <Compile Include="Configuration\Context.cs" />
