+++ test/DynamoCoreTests/Core/AssemblyLoadedEventArgsTests.cs
@@ -1,7 +1,6 @@
 ﻿using System.Reflection;
 using NUnit.Framework;
 using static Dynamo.Models.NodeModelAssemblyLoader;
-using Dynamo.Core;
 
 namespace Dynamo.Tests.Core
 {
@@ -16,24 +15,6 @@ namespace Dynamo.Tests.Core
             AssemblyLoadedEventArgs eventArgs = new AssemblyLoadedEventArgs(assemblyMock);
 
             Assert.AreEqual(assemblyMock, eventArgs.Assembly);
-        }
-    }
-
-    [TestFixture]
-    public class AssemblyConflictNotifierTest : DynamoModelTestBase
-    {
-        public override void Setup()
-        {
-            // set conflict notifier before we initialize the dynamo model (before we start loading any assemblies).
-            AssemblyConflictNotifier.SkipChecks = false;
-
-            base.Setup();
-        }
-
-        [Test]
-        public void TestAssemblyConflictNotifications()
-        {
-
         }
     }
 }
